Relocates AppEngineException and adds three new exceptions: NullSdkPathException, BadSdkLocationException and NotAFileException.
