+++ src/main/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdkAppEngineInstances.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.appengine.cloudsdk;
 
-import com.google.cloud.tools.appengine.api.AppEngineException;
+import com.google.cloud.tools.appengine.api.exceptions.AppEngineException;
 import com.google.cloud.tools.appengine.api.instances.AppEngineInstances;
 import com.google.cloud.tools.appengine.api.instances.InstancesSelectionConfiguration;
 import com.google.cloud.tools.appengine.cloudsdk.internal.args.GcloudArgs;
