+++ src/test/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdkAppEngineGenConfigTest.java
@@ -16,11 +16,9 @@
 
 package com.google.cloud.tools.appengine.cloudsdk;
 
-import com.google.cloud.tools.appengine.api.AppEngineException;
+import com.google.cloud.tools.appengine.api.exceptions.AppEngineException;
 import com.google.cloud.tools.appengine.api.genconfig.DefaultGenConfigParams;
-import com.google.cloud.tools.appengine.cloudsdk.CloudSdkAppEngineGenConfig;
 import com.google.cloud.tools.appengine.cloudsdk.internal.process.ProcessRunnerException;
-import com.google.cloud.tools.appengine.cloudsdk.CloudSdk;
 import com.google.common.collect.ImmutableList;
 
 import org.junit.Before;
